added delete to scheduling page

+++ app.py
@@ -15,21 +15,20 @@ app = Flask(__name__, static_url_path='')
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 nav = Navigation(app)
 
-#code for connection
-#MySQL Hostname
+# code for connection
+# MySQL Hostname
 app.config['MYSQL_HOST'] = 'localhost'
-#MySQL username
+# MySQL username
 app.config['MYSQL_USER'] = 'root'
-#MySQL password here in my case password is null so i left empty
+# MySQL password here in my case password is null so i left empty
 app.config['MYSQL_PASSWORD'] = 'root'
-#Database name In my case database name is projectreporting
+# Database name In my case database name is projectreporting
 app.config['MYSQL_DB'] = 'dummy_db'
 
 mysql = MySQL(app)
 
-
 # Define dict for new projects
-projects = {'AssetName': [], 'StartDate':[], 'Maintainer': [],
+projects = {'AssetName': [], 'StartDate': [], 'Maintainer': [],
             'Owner': [], 'Width': [], 'Length': [],
             'Area': [], 'Location': []}
 
@@ -45,7 +44,8 @@ nav.Bar('top', [
 def navpage():
     return render_template("navpage.html", title="Home Page")
 
-@app.route('/marketplace', methods=['GET','POST'])
+
+@app.route('/marketplace', methods=['GET', 'POST'])
 def marketplace():
     # creating variable for connection
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
@@ -54,28 +54,30 @@ def marketplace():
     # fetching all records from database
     data = cursor.fetchall()
     if request.method == "POST":
-        msg = ''
-        for getid in request.form.getlist('mycheckbox'):
-            cursor.execute("delete from components where component_id= %s", [getid])
-            mysql.connection.commit()
-            cursor.execute("select * from components")
-            data = cursor.fetchall()
-            msg = 'Successfully deleted'
-        if (not msg): msg = "There is nothing to delete"
-        return render_template("marketplace.html", msg=msg, data=data)
+        if request.form['action'] == 'Delete All Selected': #check whether form comes from delete
+            msg = ''
+            for getid in request.form.getlist('mycheckbox'):
+                cursor.execute("delete from components where component_id= %s", [getid])
+                mysql.connection.commit()
+                cursor.execute("select * from components")
+                data = cursor.fetchall()
+                msg = 'Successfully deleted'
+            if (not msg): msg = "There is nothing to delete"
+            return render_template("marketplace.html", msg=msg, data=data)
+        elif request.form['action'] == 'Show Details':
+            return render_template("component_page.html")
+
     # returning back to projectlist.html with all records from MySQL which are stored in variable data
     return render_template("marketplace.html", data=data)
 
-@app.route('/component_page.html', methods=['GET','POST'])
-def component_page():
 
+@app.route('/component_page.html', methods=['GET', 'POST'])
+def component_page():
     # creating variable for connection
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
- 
+
     cursor.execute("select * from components WHERE component_id='da26b52a-73b3-11ec-ac0f-38f9d34975e5'")
     component_data = cursor.fetchall()
-
-
 
     return render_template("component_page.html", component_data=component_data)
 
@@ -90,21 +92,22 @@ def assets_overview():
     data = cursor.fetchall()
 
     # returning back to projectlist.html with all records from MySQL which are stored in variable data
-    return render_template("assets_overview.html", data = data)
+    return render_template("assets_overview.html", data=data)
+
 
 @app.route('/add_component', methods=['GET', 'POST'])
 def add_component():
-    msg=''
-    #check to see if user is logged in otherwise, redirect user to the login page
+    msg = ''
+    # check to see if user is logged in otherwise, redirect user to the login page
     if session.get('loggedin') == True:
         msg = 'add a component'
     else:
         msg = 'You need to be logged in to add components'
         render_template('login.html')
         return redirect(url_for('login'))
-    if request.method == 'POST': #check to see if all data is filled in
+    if request.method == 'POST':  # check to see if all data is filled in
         try:
-            #create unique id for the asset
+            # create unique id for the asset
             ComponentID = uuid.uuid1()
             ComponentMaterial = request.form['ComponentMaterial']
             Category = request.form['Category']
@@ -116,22 +119,18 @@ def add_component():
             Price = request.form['Price']
             Description = request.form['comment']
 
-
-            cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)    # creating variable for connection
-
-
+            cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)  # creating variable for connection
 
             sql = "INSERT INTO components (component_id, material, category, weight, component_condition, availability, component_owner, location, price, component_description) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"
-            val = (ComponentID, ComponentMaterial, Category, Weight, Condition, Availability, Owner, Location, Price, Description)
-
-
+            val = (ComponentID, ComponentMaterial, Category, Weight, Condition, Availability, Owner, Location, Price,
+                   Description)
 
             cursor.execute(sql, val)
             mysql.connection.commit()
 
             print('Succesfull')
             flash("Project succesfully added")
-            return(redirect(url_for('marketplace.html')))
+            return (redirect(url_for('marketplace.html')))
         except:
             flash('an error occured')
 
@@ -145,7 +144,7 @@ def add_component():
 
 @app.route('/asset-components', methods=['GET', 'POST'])
 def asset_components():
-    #table
+    # table
     if request.method == 'POST':
         record_id = int(request.form['asset_id'])
         components_dataset = pd.read_excel("data/Gemeente Almere bruggen components dummy.xlsx")
@@ -157,24 +156,25 @@ def asset_components():
         cursor.execute("select * from asset_overview WHERE Assetnumber= %s", [record_id])
         bridge_dataset = cursor.fetchall()
         # bridge_dataset = components_dataset.loc[components_dataset['Assetnumber'] == record_id]
-        #image
+        # image
         img1 = os.path.join(app.config['UPLOAD_FOLDER'])
-        #uploadbutton
+        # uploadbutton
         form = MyForm()
         if form.validate_on_submit():
             filename = images.save(form.image.data)
             return f'Filename: {filename}'
 
-        #Fetch bridge specific data
+        # Fetch bridge specific data
         cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
         cursor.execute("select * from dummy_data_marketplace")
         data = cursor.fetchall()
 
         return render_template("asset_components.html", column_names=components_dataset.columns.values,
-                               row_data=list(components_dataset.values.tolist()), bridge_dataset = bridge_dataset,
-                               zip=zip, title="Asset Components", user_image = img1, form = form)
+                               row_data=list(components_dataset.values.tolist()), bridge_dataset=bridge_dataset,
+                               zip=zip, title="Asset Components", user_image=img1, form=form)
 
-@app.route('/register', methods =['GET', 'POST'])
+
+@app.route('/register', methods=['GET', 'POST'])
 def register():
     msg = ''
     if request.method == 'POST':
@@ -186,37 +186,38 @@ def register():
             companyname = request.form['companyname']
             location = str(request.form['adress'] + request.form['city'])
 
-            #connect to database
+            # connect to database
             cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-            cursor.execute('SELECT * FROM accounts WHERE email = %s', (email, ))
+            cursor.execute('SELECT * FROM accounts WHERE email = %s', (email,))
             account = cursor.fetchone()
 
             if account:
                 msg = 'Account already exists !'
             else:
 
-                cursor.execute('INSERT INTO accounts VALUES (% s, % s, % s, % s, % s, % s, NULL, NULL, NULL)', (kvknumber, phonenumber, password, email, companyname, location))
+                cursor.execute('INSERT INTO accounts VALUES (% s, % s, % s, % s, % s, % s, NULL, NULL, NULL)',
+                               (kvknumber, phonenumber, password, email, companyname, location))
 
                 mysql.connection.commit()
 
                 msg = 'You have successfully registered !'
 
-                return render_template('login.html', msg = msg)
+                return render_template('login.html', msg=msg)
         except:
             msg = 'Failed to register, please try again later'
     elif request.method == 'POST':
         msg = 'Please fill out the form !'
-    return render_template('register.html', msg = msg)
+    return render_template('register.html', msg=msg)
 
 
-@app.route('/login', methods =['GET', 'POST'])
+@app.route('/login', methods=['GET', 'POST'])
 def login():
     msg = ''
     if request.method == 'POST' and 'email' in request.form and 'password' in request.form:
         email = request.form['email']
         password = request.form['password']
         cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-        cursor.execute('SELECT * FROM accounts WHERE email = % s AND password = % s', (email, password, ))
+        cursor.execute('SELECT * FROM accounts WHERE email = % s AND password = % s', (email, password,))
         account = cursor.fetchone()
         if account:
             session['loggedin'] = True
@@ -228,55 +229,22 @@ def login():
         else:
             msg = 'Incorrect username / password !'
 
-    return render_template('login.html', msg = msg)
+    return render_template('login.html', msg=msg)
+
 
 @app.route('/logout')
 def logout():
-   session.pop('loggedin', None)
-   session.pop('email', None)
-   session.pop('companyname', None)
-   return redirect(url_for('login'))
+    session.pop('loggedin', None)
+    session.pop('email', None)
+    session.pop('companyname', None)
+    return redirect(url_for('login'))
+
 
 @app.route('/project-overview', methods=['GET', 'POST'])
 def project_overview():  # Provide forms for input
-    if request.method == 'POST': #check to see if all data is filled in
-        try:
-            #create unique id for the asset
-            UserId = session['email']
-            ProjectId = uuid.uuid1()
-            AssetName = request.form['AssetName']
-            AssetType = request.form['AssetType']
-            StartDate = request.form['start_date']
-            EndDate = request.form['end_date']
-            Maintainer = request.form['Maintainer']
-            Owner = request.form['Owner']
-            Width = request.form['Width']
-            Length = request.form['Length']
-            Location = request.form['Location']
-            ConstructionType = request.form['ConstructionType']
-
-
-            cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)    # creating variable for connection
-
-
-
-            sql = "INSERT INTO projects (user_id, project_id, assetname, assettype, startdate, enddate, maintainer, owner, width, length, location, constructiontype) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"
-            val = (UserId, ProjectId, AssetName, AssetType, StartDate, EndDate, Maintainer, Owner, Width, Length, Location, ConstructionType)
-
-
-
-            cursor.execute(sql, val)
-            mysql.connection.commit()
-
-            print('Succesfull')
-            msg = "Project succesfully added"
-            return(redirect(url_for('project_overview')))
-        except:
-            msg = 'an error occured'
-
     # creating variable for connection
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-    #collecting all the projects that the user has
+    # collecting all the projects that the user has
     if session.get('loggedin') == True:
 
         cursor.execute("select * from projects where %s = user_id", (session['email'],))
@@ -285,20 +253,66 @@ def project_overview():  # Provide forms for input
         render_template('login.html')
         return redirect(url_for('login'))
 
+    if request.method == 'POST':  # check to see if all data is filled in
+        if request.form['action'] == 'submit':  # check whether post is coming from insert
+            try:
+                # create unique id for the asset
+                UserId = session['email']
+                ProjectId = uuid.uuid1()
+                AssetName = request.form['AssetName']
+                AssetType = request.form['AssetType']
+                StartDate = request.form['start_date']
+                EndDate = request.form['end_date']
+                Maintainer = request.form['Maintainer']
+                Owner = request.form['Owner']
+                Width = request.form['Width']
+                Length = request.form['Length']
+                Location = request.form['Location']
+                ConstructionType = request.form['ConstructionType']
+
+                cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)  # creating variable for connection
+
+                sql = "INSERT INTO projects (user_id, project_id, assetname, assettype, startdate, enddate, maintainer, owner, width, length, location, constructiontype) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"
+                val = (
+                UserId, ProjectId, AssetName, AssetType, StartDate, EndDate, Maintainer, Owner, Width, Length, Location,
+                ConstructionType)
+
+                cursor.execute(sql, val)
+                mysql.connection.commit()
+
+                print('Succesfull')
+                msg = "Project succesfully added"
+                return (redirect(url_for('project_overview')))
+            except:
+                msg = 'an error occurred'
+        elif request.form['action'] == 'Delete All Selected':  # check whether post is coming from delete
+            msg = ''
+            for getid in request.form.getlist('mycheckbox'):
+                cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
+                cursor.execute("delete from projects where project_id= %s", [getid])
+                mysql.connection.commit()
+                cursor.execute("select * from projects where %s = user_id", (session['email'],))
+                project_data = cursor.fetchall()
+                msg = 'Successfully deleted'
+            if (not msg): msg = "There is nothing to delete"
+            return render_template("project_overview.html", msg=msg, projects_df=project_data)
+        elif request.form['action'] == 'Show Details':  #check whether post is coming from 'show details'
+            return render_template("component_page.html")
+
     # Get possible construction and asset types from Asset database and convert it to a dataframe
     cursor.execute("select * from asset_overview")
     asset_data = pd.DataFrame(cursor.fetchall())
 
-    #reindex the dataframe and get the right information
+    # reindex the dataframe and get the right information
     dataset = asset_data.reindex(columns=asset_data.columns.tolist() + ['Open_Asset'])
     construction_type = dataset['ConstructionType'].unique()
     asset_type = dataset['AssetType'].unique()
-
 
     return render_template("project_overview.html",
                            projects_df=project_data,
                            construction_type=construction_type,
                            asset_type=asset_type)
+
 
 ## IMPORT image
 app.config['SECRET_KEY'] = 'thisisasecret'
@@ -306,8 +320,10 @@ app.config['UPLOADED_IMAGES_DEST'] = 'static/uploads'
 images = UploadSet('images', IMAGES)
 configure_uploads(app, images)
 
+
 class MyForm(FlaskForm):
     image = FileField('image')
+
 
 # run the application
 if __name__ == '__main__':
